refactor: create abstractions
